backend/requirements.txt
azure-core==1.35.0
azure-storage-blob==12.25.1
beanie==1.26.0
email-validator
fastapi==0.104.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn[standard]==0.24.0
